+++ src/event-listeners-handler/EventListenersHandler.test.ts
@@ -2,23 +2,16 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 import EventListenersHandler from './EventListenersHandler'
 import ClickEventListener from '../event-listeners/ClickEventListener'
 import BeforeUnloadEventListener from '../event-listeners/BeforeUnloadEventListener'
-import MemoryUserActionsHistory from '../user-actions-history/MemoryUserActionsHistory'
-import UserActionHandler from '../user-action/UserActionHandler'
-import { nop } from '../utils/nop'
-import UserActionsHistory from '../user-actions-history/UserActionsHistory'
-import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler'
+import IUserActionHandler, { NopUserActionHandler } from '../user-action/IUserActionHandler'
 
 describe('EventListenersHandler', () => {
   let eventListenersHandler: EventListenersHandler
   let clickEventListener: ClickEventListener
   let beforeUnloadEventListener: BeforeUnloadEventListener
-  let userActionsHistory: UserActionsHistory
-  let userActionHandler: UserActionHandler
+  let userActionHandler: IUserActionHandler
 
   beforeEach(() => {
-    userActionsHistory = new MemoryUserActionsHistory()
-    const sessionIdHandler = new MemorySessionIdHandler(() => 'aaa-111', 500)
-    userActionHandler = new UserActionHandler(sessionIdHandler, 0, nop, nop, userActionsHistory)
+    userActionHandler = new NopUserActionHandler()
     clickEventListener = new ClickEventListener(userActionHandler, window)
     beforeUnloadEventListener = new BeforeUnloadEventListener(userActionHandler, window)
     eventListenersHandler = new EventListenersHandler([clickEventListener, beforeUnloadEventListener])