+++ src/event-listeners/ClickEventListener.ts
@@ -1,10 +1,10 @@
-import UserActionHandler from '../user-action/UserActionHandler'
 import { createTargetedUserAction } from '../user-action/createTargetedUserAction'
 import { UserActionType } from '../types'
 import TargetedEventListener, { TargetEventListenerOptions } from './TargetedEventListener'
+import IUserActionHandler from '../user-action/IUserActionHandler'
 
 class ClickEventListener extends TargetedEventListener {
-  constructor(userActionHandler: UserActionHandler, window: Window, options: TargetEventListenerOptions = {}) {
+  constructor(userActionHandler: IUserActionHandler, window: Window, options: TargetEventListenerOptions = {}) {
     super(userActionHandler, UserActionType.Click, window, options)
   }