+++ CHANGELOG.md
@@ -21,7 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.7.1...main)
+## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.8.1...main)
 
 ### Info
 
@@ -34,6 +34,18 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 
 ### Fixed
+
+## [3.8.1](https://github.com/Smartesting/gravity-data-collector/compare/v3.8.0...v3.8.1)
+
+### Changed
+
+- Cypress event tracking: skip some events (named 'then', 'task' or 'wrap')
+
+## [3.8.0](https://github.com/Smartesting/gravity-data-collector/compare/v3.7.1...v3.8.0)
+
+### Added
+
+- track Cypress test information (if available) as a new `TestCommand` user action
 
 ## [3.7.1](https://github.com/Smartesting/gravity-data-collector/compare/v3.7.0...v3.7.1)